Lite.Conformance/overrides/lite/domcontentloaded.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<title>Lite DOMContentLoaded: firing order, readyState, event shape</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
// In-position script: readyState is still "loading" and DOMContentLoaded has not fired yet,
// so listeners registered here must fire later (after deferred scripts, before load).
var order = [];

test(function () {
  assert_equals(document.readyState, 'loading');
}, "readyState is 'loading' during in-position script execution");

async_test(function (t) {
  document.addEventListener('DOMContentLoaded', t.step_func_done(function (e) {
    order.push('dcl-document');
    assert_equals(e.type, 'DOMContentLoaded', "event type is case-sensitive");
    assert_true(e.bubbles, "DOMContentLoaded bubbles");
    assert_equals(document.readyState, 'interactive', "readyState is 'interactive' inside DOMContentLoaded");
  }));
}, "document DOMContentLoaded listener fires with the spec event shape");

async_test(function (t) {
  window.addEventListener('DOMContentLoaded', t.step_func_done(function () {
    order.push('dcl-window');
  }));
}, "window DOMContentLoaded listener fires");

async_test(function (t) {
  window.addEventListener('load', t.step_func_done(function () {
    order.push('load');
    assert_equals(document.readyState, 'complete', "readyState is 'complete' inside load");
    assert_equals(order.indexOf('dcl-document'), 0, "document DOMContentLoaded fired first");
    assert_true(order.indexOf('dcl-window') < order.indexOf('load'),
                "window DOMContentLoaded fired before load");
  }));
}, "load fires after DOMContentLoaded with readyState 'complete'");
</script>
</body>
</html>
